# swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Sesi",
    "description": "Documentação da API do projeto Sesi (demonstração de segurança).\n\n**GET** = buscar dados | **POST** = criar/enviar | **DELETE** = remover",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Backend local"
    }
  ],
...
  "paths": {
...
    "/api/usuarios/{id}": {
      "get": {
        "tags": ["Usuários"],
        "summary": "Buscar um usuário por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "example": "1779310561032"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado (sem senha)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "usuario": { "$ref": "#/components/schemas/UsuarioPublico" }
                  }
                }
              }
            }
          },
          "404": { "description": "Não encontrado" }
        }
      }
    },
    "/api/usuarios/{id}/moedas": {
      "post": {
        "tags": ["Usuários"],
        "summary": "Adicionar moedas ao usuário",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantidade": {
                    "type": "number",
                    "default": 100,
                    "example": 100
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Moedas adicionadas" },
          "404": { "description": "Usuário não encontrado" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UsuarioPublico": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "nome": { "type": "string" },
          "email": { "type": "string" },
          "cpf": { "type": "string" },
          "moedas": { "type": "number" }
        }
      },
...
    }
  }
}